Extract mutation & request utility from middleware
In our calling code we have a need to persist a LeadSource
manually when a user is not authenticated and we know this
can be considered reliable.

But before this change, we would have had to write custom
code rather than importing functions from this library as
the "working logic" was hidden inside middleware methods.

This PR moves two things outside of the middleware, so that
they can be called from elsewhere - and updates the middleware
to make that same call.

1) The "has the request got the utm params we want?" utility.

2) The "persist the lead source and clear the session" function.

Updates so make user passing explicit